+++ src/main/java/org/topj/methods/request/Transfer.java
@@ -64,7 +64,7 @@ public class Transfer implements Request {
             XTransaction xTransaction = new XTransaction();
             xTransaction.setTransactionType(XTransactionType.Transfer);
             xTransaction.setLastTransNonce(account.getNonce());
-            xTransaction.setFireTimestamp(new Date().getTime()/1000);
+            xTransaction.setFireTimestamp(BigInteger.valueOf(new Date().getTime()/1000));
 
             xTransaction.setExpireDuration(TopjConfig.getExpireDuration());
             xTransaction.setLastTransHash(account.getLastHashXxhash64());
